+++ SeaPublicWebsite/ExternalServices/BreApi.cs
@@ -1,0 +1,7 @@
+﻿namespace SeaPublicWebsite.ExternalServices
+{
+    public class BreApi
+    {
+    
+    }
+}
